Implement rewrite logic for multi-part tables
