+++ sources/sql/src/rewrite/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod ast;
+pub(crate) mod plan;
